requirements.txt
Django==2.1.4
djangorestframework==3.9.0
mysqlclient==1.3.12
pytz==2018.7
